Rename lib folder to public API
