+++ lib/public-api/card.class.ts
