+++ lib/public-api/interfaces.ts
