+++ lib/public-api/mapProperties.ts
